requirements.txt
django>=3.0.0
djangorestframework>=3.11.0
environs
loguru
tornado
httpx
requests